requirements.txt
flask==3.1.2
flask-login==0.6.3
flask-wtf==1.2.2
gunicorn==23.0.0
matplotlib==3.10.7
openpyxl==3.1.5
pandas==2.3.3
psycopg2-binary==2.9.10
reportlab==4.2.5
requests==2.32.5
sqlalchemy==2.0.44
werkzeug==3.1.4
